requirements-py314.txt
# === Core (Python 3.14 Compatible) ===
pandas>=2.0
numpy>=1.26
geopandas>=0.14
shapely>=2.0
h3>=4.0
pyarrow>=14.0

# === Data Fetching ===
requests>=2.31
osmnx>=1.7
overpy>=0.7

# === Weather ===
meteostat>=1.6

# === Computer Vision ===
torch>=2.1
torchvision>=0.16
Pillow>=10.0
rasterio>=1.3
earthengine-api>=0.1.370

# === Modeling ===
xgboost>=2.0
scikit-learn>=1.3
shap>=0.43

# === Experiment Tracking ===
mlflow>=2.9

# === Visualization & Dashboard ===
plotly>=5.18
streamlit>=1.29
folium>=0.15
streamlit-folium>=0.15
matplotlib>=3.8
seaborn>=0.13

# === Ethics / Fairness ===
fairlearn>=0.9

# === Utilities ===
python-dotenv>=1.0
tqdm>=4.66
joblib>=1.3
pyngrok>=7.0
